requirements.txt
streamlit
pandas
Pillow
GitPython
